paperai/src/paperai/data_loader.py
```python
import logging

# Initialize logger
logger = logging.getLogger(__name__)

class DataLoader:
    def __init__(self):
        pass

    def load_data(self, data):
        """Loads and preprocesses data, handling null and empty strings."""
        processed_data = []
        for item in data:
            processed_item = {}
            for key, value in item.items():
                if value is None:
                    logger.warning(f"Null value encountered for key: {key}. Replacing with empty string.")
                    processed_item[key] = ""  # Replace null with empty string
                elif isinstance(value, str) and value.strip() == "":
                    logger.warning(f"Empty string encountered for key: {key}. Replacing with empty string.")
                    processed_item[key] = ""  # Replace empty string with empty string
                else:
                    processed_item[key] = value
            processed_data.append(processed_item)

        return processed_data


# Example usage (you might have this in a separate script or as part of cli.py)
if __name__ == '__main__':
    # Configure logging
    logging.basicConfig(level=logging.WARNING)

    # Create dummy data with null and empty strings
    dummy_data = [
        {"title": "Paper 1", "author": "John Doe", "abstract": None},
        {"title": "Paper 2", "author": "", "abstract": "This is an abstract."},
        {"title": "Paper 3", "author": "Jane Smith", "abstract": "  "},
        {"title": "Paper 4", "author": "", "abstract": None},
        {"title": "Paper 5", "author": "Peter Jones", "abstract": "Valid abstract"}
    ]

    # Load the data
    data_loader = DataLoader()
    loaded_data = data_loader.load_data(dummy_data)

    # Print the processed data
    print(loaded_data)
```